them pid nhung chua chay

+++ 1_MainFunctions/3_SensorFunctions.ino
@@ -1,6 +1,6 @@
 // Function:
 /*Get obstacle distance from ultrasonic sensor*/
-void GetObstacleDistance(int trigPin, int echoPin, float& distance)
+void GetObstacleDistance(int trigPin, int echoPin, double& distance)
 {
   unsigned long duration;
   /*Sets the trigPin states*/
